requirements.txt
# streamlit
# langchain
# ipykernel
# python-dotenv
# langchain_community
# pypdf
# bs4
# langchain-text-splitters
# sentence_transformers
# langchain_huggingface
# faiss-cpu
# chromadb
# langchain-groq
# langchain-ollama
# streamlit
# chromadb

streamlit
langchain
langchain-groq
langchain-ollama
chromadb
faiss-cpu
sentence_transformers
langchain-huggingface
python-dotenv
pypdf
bs4
ipykernel